# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Fiuumber User API",
        "description": "CRUD",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [{
            "url": "https://fiuumber-api-users.herokuapp.com/api/users-service",
            "description": "Heroku"
        },
        {
            "url": "http://localhost:8081/api/users-service",
            "description": "Localhost in port 8081"
        }
    ],
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/user/{id}": {
            "parameters": [{
                "in": "path",
                "name": "id",
                "required": true,
                "description": "User id",
                "schema": {
                    "type": "integer",
                    "example": 1
                }
            }],
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user with given ID",
                "parameters": [{
                    "in": "path",
                    "name": "id",
                    "required": true,
                    "description": "User id",
                    "schema": {
                        "type": "integer",
                        "example": 1
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "404": {
                        "description": "Failed. User not found."
                    }
                }
            },
            "delete": {
                "summary": "Delete user with given ID",
                "tags": [
                    "User"
                ],
                "parameters": [{
                    "in": "path",
                    "name": "id",
                    "required": true,
                    "description": "Passenger id",
                    "schema": {
                        "type": "integer",
                        "example": 1
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    "404": {
                        "description": "Failed. User not found."
                    }
                }
            }
        },
...
    },
    "definitions": {
        "id": {
            "properties": {
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "User": {
            "type": "object",
            "properties": {
                "userId": {
                    "type": "number",
                    "example": 1
                },
                "email": {
                    "type": "string",
                    "example": "acai@fi.uba.ar"
                },
                "firstName": {
                    "type": "boolean",
                    "example": "Ana"
                },
                "lastName": {
                    "type": "string",
                    "example": "Cai"
                },
                "address": {
                    "type": "string",
                    "example": "Tucuman 111"
                },
                "password": {
                    "type": "string",
                    "example": "fasbf62823hd91d019dj12j1209j09j90"
                },
                "username": {
                    "type": "string",
                    "example": "anichucai"
                },
                "wallet": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string",
                            "example": "111"
                        },
                        "walletPrivateKey": {
                            "type": "string",
                            "example": "1dsnh2337d82hdu2@$%@!%tewdfwdww"
                        }
                    }
                },
                "blocked": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
    }
}